# static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vendor Data Quality & Trust Issues Platform API",
    "description": "Production-grade RESTful API documentation for Vendor Risk Assessment, Fraud Detection, Compliance Auditing, AI Copilot, and Real-Time Telemetry.",
    "version": "2.0.0",
    "contact": {
      "name": "Enterprise Security & API Team",
      "email": "api-security@vendor-trust.local"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    },
    {
      "url": "https://localhost",
      "description": "Production Nginx Reverse Proxy"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token obtained from `/api/v2/auth/login`"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string", "example": "Unauthorized access or invalid token." },
          "code": { "type": "integer", "example": 401 }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "example": "admin@vendor.com" },
          "password": { "type": "string", "example": "AdminPass123!" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "token": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..." },
          "user": {
            "type": "object",
            "properties": {
              "id": { "type": "integer", "example": 1 },
              "email": { "type": "string", "example": "admin@vendor.com" },
              "role": { "type": "string", "example": "Admin" }
            }
          }
        }
      },
...
    }
  },
  "paths": {
    "/api/v2/auth/login": {
      "post": {
        "summary": "Authenticate User & Issue JWT Token",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}